requirements-dev.txt
# Development-only dependencies — not needed at runtime or in the Docker image.
# Install with: pip install -r requirements-dev.txt

# Testing
pytest==8.2.0
pytest-cov==5.0.0         # coverage reports

# Linting
ruff==0.4.4               # fast linter + formatter (replaces flake8, isort)

# Type checking
mypy==1.10.0
pandas-stubs==2.2.1.240316  # type stubs for pandas
